Revert "Prevent unsafe narrowing: base/*/*"
This reverts commit 597a40586dfc5e8078fef18200feb594b422e924.

Reason for revert: Caused failing Linux MSan tests

https://ci.chromium.org/p/chromium/builders/ci/Linux%20ChromiumOS%20MSan%20Tests

Manually verified by:
1) running MSan tests on tip-of-tree (tests failed).
2) reverting r1009211 (tests passed).
3) reverting revert in step 2 (test failures returned).

Original change's description:
> Prevent unsafe narrowing: base/*/*
>
> * Change to more correct types
> * Add some static_casts where safe
>
> Bug: 1292951
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3673877
> Commit-Queue: Peter Kasting <pkasting@chromium.org>
> Auto-Submit: Peter Kasting <pkasting@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1009211}

Bug: 1292951
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3681982
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Commit-Queue: Chris Mumford <cmumford@google.com>
Owners-Override: Chris Mumford <cmumford@google.com>
Cr-Commit-Position: refs/heads/main@{#1009381}

+++ base/allocator/partition_allocator/starscan/state_bitmap.h
@@ -404,7 +404,7 @@ StateBitmap<PageSize, PageAlignment, AllocationAlignment>::IterateImpl(
     CellType value = LoadCell(cell_index);
     while (value) {
       const size_t trailing_zeroes =
-          static_cast<size_t>(base::bits::CountTrailingZeroBits(value) & ~0b1);
+          (base::bits::CountTrailingZeroBits(value) & ~0b1);
       const size_t clear_value_mask =
           ~(static_cast<CellType>(kStateMask) << trailing_zeroes);
       const CellType bits = (value >> trailing_zeroes) & kStateMask;
